remove LICENSE.txt and setup.cfg file from root
The install creates LICENSE.txt and setup.cfg files in the root of the
package which can conflict with other packages.  The setup.cfg file
especially can cause issues since it contains setting for pytest which
will be picked up in any directory below the root directory, for example
conda-bld environments.